Change default exporter source_ref to "stable"
This will address the core issue we saw in
https://github.com/konveyor/pelorus/issues/303, where pinning your repo to a
specific tag still has the exporters pointed at `master`--
causing things like python code versions and container versions to get out of
sync.
Making `stable` the default will make for a better user experience,
especially for updates.

I've created the
[stable branch](https://github.com/konveyor/pelorus/tree/stable)
which right now points to the same commit as the
[v1.3.0](https://github.com/konveyor/pelorus/tree/v1.3.0) tag.