Improves workflow for building Pelorus images
Related #63 and #358

Cleans up and improves workflow for building Pelorus images
using GitHub actions. Main difference is to set correctly tags
during build process rather then re-tagging them after they
are built and moving number of common values to variables.

On top of that the image namespace is not included in the
action itself, which easy testing of build actions in the
forked repositories.