Fix Issue #446 where committime exporter had to be restarted (#461)
The issue with commit time exporter is that new namespace could
not be registered because object created at the very first step
by GitFactory assigned watched namespaces and they were not
updated during lifetime of the application. Similar problem
occured when there were no NAMESPACES passed to the application
in which case object was assigned with all namespaces, but never
freed. This change fixes the issue where we don't associate
instance variable and gather this value every time collector
is generating metrics.