Nit change to inform that exporter is ran from image.
When the exporter is ran from the pre-built image, no
OPENSHIFT_BUILT_* env variables are available. There is also
no information accessible from the image itself about the
source code from which the image was built. This is fine
as we have knowledge about the image sha and we can easily
identify the source code for it.

To not confuse users a small nit to the message change is required.