Small release bump - 1.7.1 (#601)
We need to bump to 1.7.1 from just released 1.7.0. This is because
the Draft release was not published after that minor bump and
followed by other PRs that did not create new quay images #600.

Note that exporter quay images are being built not on every PR
merge, just the one which is touching exporter files or _buildconfig.

This is corner case where 1.7.0 tag was trying to find latest
hash from the github, however quay was missing that hash due to
image not being built.

In the first place the publish release was supposed to be
created after minor release PR was merged #599.