WIP: Allow user to override github bug label
Issues in github for personal repos by default have the label "bug".
It is possible for an org or user to use any label to indicate a bug.
For example OpenShift changed the label "bug" to "kind/bug"

This change is meant to make the label used for issues only in Github.
Allow the user to override the label.