Update tekton to label committime on image for s2itekton
tekton s2i task provides no builds, buildConfig etc. The
only artifact is an image.  Enhance the labels2i tekton task
to also annotate w/ the commit time.

When creating a result in tekton for the commit time the result
has a new line.  Use strip() on the annotation label.