Adds continous metric to the deploytime
Change that adds to the deploytime continous time series for the active
deployments.

Deployment with a timestamp creates an event, which may be in the past.
Having just an event we are unable to make calculations based on the
currently active deployments, because one may be gone from the cluster.

Adding deployment_active metric to already existing deploy_timestamp
will allow to better visualize Lead Time for Change, where the
"dots" are connected to each other, plus in the future we may be
able to answer additional questions such as how many deployed apps
are still working.

On top of that this change ensures one metric is presented
only one time in the series rather then multiple times. The metrics
were gathered several times multiplied by the number of found
unique deployments.

Signed-off-by: Michal Pryc <mpryc@redhat.com>

Clean up deploytime

- Simplifies reporting deploy_timestamp and deployment_active
- Simplifies DeployTimeMetric creation (datetime handling)

Signed-off-by: Kevin M Granger <kgranger@redhat.com>
Co-authored-by: Kevin M Granger <kgranger@redhat.com>