Overhaul committime collector(s).
There were several typing issues and unclear APIs in the committime exporter.

CommitMetric was used as a big blob.
It carried all inputs and outputs, with almost no guarantees for
when and where things would be set. It also had several unused fields.

The way it was meant to be handled by implementors was also unclear:

do they modify it on the passed in CommitMetric?
do they return a CommitMetric with the committime updated?
what if they return None?
what if the information needed is missing?
There are several spots where a type checker would complain.
Including some spots where the variable being used could be unbound!

Signed-off-by: Kevin M Granger <kgranger@redhat.com>