New layout for the exporter docs
Current layout for the Pelorus Exporters docs causes the page
to be very long. Split into per exporter configuration make it
easier to focus on a particular part.

Also we don't need to inform user about different exporter
installation paths (source code, quay images, hashes etc).
This is more of an development doc, so moving that part to the
Development section.

Signed-off-by: Michal Pryc <mpryc@redhat.com>