Add secret to secure payload for the webhook exporter
Adds option to configure webhook exporter to share SECRET_TOKEN
to validate payload integrity.

With the webhook configured to use SECRET_TOKEN the sender is
obligated to calculate the SHA256 hash of the payload and include
that hash as a header during POST method, otherwise the payload
will not be received and recorded by the webhook.

Signed-off-by: Michal Pryc <mpryc@redhat.com>