Pelorus-Operator 0.0.6 with fresh patches and updated release scripts
This change includes Pelorus-Operator 0.0.6 which refreshes patches,
updated linting of the charts so they now must match the pelorus-operator
charts.

Also it consists of change where failing patches were allowing to proceed
with pelorus-operator creation.

To create release it's enough to type: make release

The RC releases are still two step as the create_pelorus_operator must
have explicit rc version to be passed for the operator rc version.

Signed-off-by: Michal Pryc <mpryc@redhat.com>