Add the timestamp to the webhook exporter.
Currently a valid event is the one that happened within 30min from the
time of sending that event to our webhook exporter. This is the
limitation of Prometheus that is not accepting events that
are too old or are too far into the future.

Fixes also webhook exporter to contain timestamps for the events
that occured.

Signed-off-by: Michal Pryc <mpryc@redhat.com>