Rework of deploytime codebase
Move functions to the common openshift library.

Optimize querying for the parent objects such as ReplicaSet
and ReplicationController. With this implementation we relay on known
api versions and api objects directly from the pod.

Added cache to optimize queries to the OpenShift API for the
parent objects. For the running pods we do want to always have
the latest state, which means we can't do caching there.

Reworked extracting image SHAs from the running pod instances,
which is a base for the skopeo as we also get the full URI of the
image.

Signed-off-by: Michal Pryc <mpryc@redhat.com>