Add Timestamp to the Failure exporter
Adding Timestamp to the Failure exporter is important to properly calculate
Change Failure Rate Over Time (CFR)

The CFR metric is a simple number that represents the number of failed deployments
to the number of all deployments. As such without the timestamp we were adding the
Failures to the previously known ones and never calculating properly based on a chosen
time window.

The max_over_time is required to be able to deduct creation time of the failure from the
resolution time of that failure to properly calculate Mean Time to Restore and other metrics.
Without that we would almost never had the time window that both metrics create/failure would
be within it and as such the deduction would return nothing.

Signed-off-by: Michal Pryc <mpryc@redhat.com>